Super glitchy shader implementation of compression codec.
Committing for posterity; maybe this can be massagged into some kind of working implementation with a lot of work but I just don't feel like it. Not when the compression savings still won't let it approach a proper streaming video codec.

Maybe I should just bite the bullet and really try to get some kind of webrtc router implementation running. -_-